Timeline Added
A timeline function was added to so that on the google sheet you can include when a location was created and when it ended. Then you can scroll through to see how locations change over time.

Also updated the url so that these parameters can be fed in such as location, and time for hyperlinking.